requirements.txt
attrs==23.1.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
DateTime==5.2
distlib==0.3.9
filelock==3.16.1
Flask==3.1.0
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
MarkupSafe==3.0.2
mysqlclient==2.2.7
numpy==1.25.2
opencv-python==4.8.0.76
pillow==11.2.1
platformdirs==4.3.6
pyarrow==13.0.0
pylance==0.8.5
pyproject-toml==0.0.10
python-dotenv==1.1.0
pytz==2023.3.post1
PyYAML==6.0.2
referencing==0.30.2
render==1.0.0
rpds-py==0.10.3
SQLAlchemy==2.0.41
template==0.7.6
toml==0.10.2
typing_extensions==4.13.2
virtualenv==20.28.1
Werkzeug==3.1.3
wtf==0.1
zope.interface==6.0
